layer roles function moved to utils
