+++ tests/models/abiotic_simple/test_abiotic_simple_model.py
@@ -16,23 +16,10 @@ from virtual_rainforest.models.abiotic_simple.abiotic_simple_model import (
 )
 
 
-def test_set_layer_roles():
-    """Test correct order of layers."""
-    from virtual_rainforest.models.abiotic_simple.abiotic_simple_model import (
-        set_layer_roles,
-    )
-
-    assert set_layer_roles(10, 2) == (
-        ["above"] + ["canopy"] * 10 + ["subcanopy"] + ["surface"] + ["soil"] * 2
-    )
-
-
 @pytest.mark.parametrize(
-    "soil_layers,canopy_layers,ini_soil_moisture,raises,expected_log_entries",
+    "ini_soil_moisture,raises,expected_log_entries",
     [
         (
-            2,
-            10,
             50.0,
             does_not_raise(),
             (
@@ -74,56 +61,6 @@ def test_set_layer_roles():
             ),
         ),
         (
-            -2,
-            10,
-            50.0,
-            pytest.raises(InitialisationError),
-            (
-                (
-                    ERROR,
-                    "There has to be at least one soil layer in the abiotic model!",
-                ),
-            ),
-        ),
-        (
-            2,
-            -3,
-            50.0,
-            pytest.raises(InitialisationError),
-            (
-                (
-                    ERROR,
-                    "There has to be at least one canopy layer in the abiotic model!",
-                ),
-            ),
-        ),
-        (
-            2.5,
-            10,
-            50.0,
-            pytest.raises(InitialisationError),
-            (
-                (
-                    ERROR,
-                    "The number of soil layers must be an integer!",
-                ),
-            ),
-        ),
-        (
-            2,
-            3.4,
-            50.0,
-            pytest.raises(InitialisationError),
-            (
-                (
-                    ERROR,
-                    "The number of canopy layers must be an integer!",
-                ),
-            ),
-        ),
-        (
-            2,
-            10,
             -50.0,
             pytest.raises(InitialisationError),
             (
@@ -134,8 +71,6 @@ def test_set_layer_roles():
             ),
         ),
         (
-            2,
-            10,
             DataArray([50, 30, 20]),
             pytest.raises(InitialisationError),
             (
@@ -150,12 +85,12 @@ def test_set_layer_roles():
 def test_abiotic_simple_model_initialization(
     caplog,
     dummy_climate_data,
-    soil_layers,
-    canopy_layers,
     ini_soil_moisture,
     raises,
     expected_log_entries,
     layer_roles_fixture,
+    soil_layers=2,
+    canopy_layers=10,
 ):
     """Test `AbioticSimpleModel` initialization."""
