+++ virtual_rainforest/models/abiotic_simple/abiotic_simple_model.py
@@ -18,7 +18,7 @@ that all model configuration failures can be reported as one.
 
 from __future__ import annotations
 
-from typing import Any, List
+from typing import Any
 
 import numpy as np
 from pint import Quantity
@@ -28,6 +28,7 @@ from virtual_rainforest.core.base_model import BaseModel
 from virtual_rainforest.core.data import Data
 from virtual_rainforest.core.exceptions import InitialisationError
 from virtual_rainforest.core.logger import LOGGER
+from virtual_rainforest.core.utils import set_layer_roles
 from virtual_rainforest.models.abiotic_simple import simple_regression
 
 
@@ -69,35 +70,6 @@ class AbioticSimpleModel(BaseModel):
         initial_soil_moisture: float,
         **kwargs: Any,
     ):
-        # sanity checks for soil and canopy layers
-        if soil_layers < 1:
-            to_raise = InitialisationError(
-                "There has to be at least one soil layer in the abiotic model!"
-            )
-            LOGGER.error(to_raise)
-            raise to_raise
-
-        if not isinstance(soil_layers, int):
-            to_raise = InitialisationError(
-                "The number of soil layers must be an integer!"
-            )
-            LOGGER.error(to_raise)
-            raise to_raise
-
-        if canopy_layers < 1:
-            to_raise = InitialisationError(
-                "There has to be at least one canopy layer in the abiotic model!"
-            )
-            LOGGER.error(to_raise)
-            raise to_raise
-
-        if canopy_layers != int(canopy_layers):
-            to_raise = InitialisationError(
-                "The number of canopy layers must be an integer!"
-            )
-            LOGGER.error(to_raise)
-            raise to_raise
-
         # sanity checks for initial soil moisture
         if type(initial_soil_moisture) is not float:
             to_raise = InitialisationError("The initial soil moisture must be a float!")
@@ -217,32 +189,3 @@ class AbioticSimpleModel(BaseModel):
 
     def cleanup(self) -> None:
         """Placeholder function for abiotic model cleanup."""
-
-
-def set_layer_roles(canopy_layers: int, soil_layers: int) -> List[str]:
-    """Create a list of layer roles.
-
-    This function creates a list of layer roles for the vertical dimension of the
-    Virtual Rainforest. The layer above the canopy is defined as 0 (canopy height + 2m)
-    and the index increases towards the bottom of the soil column. The canopy includes
-    a maximum number of canopy layers (defined in config) which are filled from the top
-    with canopy node heights from the plant module (the rest is set to NaN). Below the
-    canopy, we currently set one subcanopy layer (around 1.5m above ground) and one
-    surface layer (0.1 m above ground). Below ground, we include a maximum number of
-    soil layers (defined in config); the deepest layer is currently set to 1 m as the
-    temperature there is fairly constant and equals the mean annual temperature.
-
-    Args:
-        canopy_layers: number of canopy layers
-        soil_layers: number of soil layers
-
-    Returns:
-        List of canopy layer roles
-    """
-    return (
-        ["above"]
-        + ["canopy"] * canopy_layers
-        + ["subcanopy"]
-        + ["surface"]
-        + ["soil"] * soil_layers
-    )
